requirements.txt
python-telegram-bot==22.0
yt-dlp